+++ zcash_client_backend/src/data_api/wallet.rs
@@ -1,16 +1,11 @@
 use std::num::NonZeroU32;
 
-use shardtree::{error::ShardTreeError, store::ShardStore, ShardTree};
-
 use zcash_primitives::{
-    consensus::{self, BlockHeight, NetworkUpgrade},
+    consensus::{self, NetworkUpgrade},
     memo::MemoBytes,
     sapling::{
-        self,
         note_encryption::{try_sapling_note_decryption, PreparedIncomingViewingKey},
         prover::{OutputProver, SpendProver},
-        zip32::DiversifiableFullViewingKey,
-        Node,
     },
     transaction::{
         builder::Builder,
@@ -26,12 +21,11 @@ use crate::{
     data_api::{
         error::Error, wallet::input_selection::Proposal, DecryptedTransaction, SentTransaction,
         SentTransactionOutput, WalletCommitmentTrees, WalletRead, WalletWrite,
-        SAPLING_SHARD_HEIGHT,
     },
     decrypt_transaction,
     fees::{self, ChangeValue, DustOutputPolicy},
     keys::UnifiedSpendingKey,
-    wallet::{NoteId, OvkPolicy, ReceivedSaplingNote, Recipient},
+    wallet::{OvkPolicy, Recipient, WalletNote},
     zip321::{self, Payment},
     PoolType, ShieldedProtocol,
 };
@@ -591,26 +585,24 @@ where
     if let Some(sapling_inputs) = proposal.sapling_inputs() {
         wallet_db.with_sapling_tree_mut::<_, _, Error<_, _, _, _>>(|sapling_tree| {
             for selected in sapling_inputs.notes() {
-                let (note, scope, merkle_path) = select_key_for_note(
-                    sapling_tree,
-                    selected,
-                    &dfvk,
-                    sapling_inputs.anchor_height(),
-                )?
-                .ok_or_else(|| {
-                    Error::NoteMismatch(NoteId::new(
-                        *selected.txid(),
-                        ShieldedProtocol::Sapling,
-                        selected.output_index(),
-                    ))
-                })?;
+                match selected.note() {
+                    WalletNote::Sapling(note) => {
+                        let key = match selected.spending_key_scope() {
+                            Scope::External => usk.sapling().clone(),
+                            Scope::Internal => usk.sapling().derive_internal(),
+                        };
 
-                let key = match scope {
-                    Scope::External => usk.sapling().clone(),
-                    Scope::Internal => usk.sapling().derive_internal(),
-                };
+                        let merkle_path = sapling_tree.witness_at_checkpoint_id_caching(
+                            selected.note_commitment_tree_position(),
+                            &sapling_inputs.anchor_height(),
+                        )?;
 
-                builder.add_sapling_spend(key, note, merkle_path)?;
+                        builder.add_sapling_spend(key, note.clone(), merkle_path)?;
+                    }
+                    WalletNote::Orchard(_) => {
+                        panic!("Orchard spends are not yet supported");
+                    }
+                }
             }
             Ok(())
         })?;
@@ -880,41 +872,4 @@ where
         OvkPolicy::Sender,
         &proposal,
     )
-}
-
-#[allow(clippy::type_complexity)]
-fn select_key_for_note<N, S: ShardStore<H = Node, CheckpointId = BlockHeight>>(
-    commitment_tree: &mut ShardTree<
-        S,
-        { sapling::NOTE_COMMITMENT_TREE_DEPTH },
-        SAPLING_SHARD_HEIGHT,
-    >,
-    selected: &ReceivedSaplingNote<N>,
-    dfvk: &DiversifiableFullViewingKey,
-    anchor_height: BlockHeight,
-) -> Result<Option<(sapling::Note, Scope, sapling::MerklePath)>, ShardTreeError<S::Error>> {
-    // Attempt to reconstruct the note being spent using both the internal and external dfvks
-    // corresponding to the unified spending key, checking against the witness we are using
-    // to spend the note that we've used the correct key.
-    let external_note = dfvk
-        .diversified_address(selected.diversifier())
-        .map(|addr| addr.create_note(selected.value().try_into().unwrap(), selected.rseed()));
-    let internal_note = dfvk
-        .diversified_change_address(selected.diversifier())
-        .map(|addr| addr.create_note(selected.value().try_into().unwrap(), selected.rseed()));
-
-    let expected_root = commitment_tree.root_at_checkpoint_id(&anchor_height)?;
-    let merkle_path = commitment_tree.witness_at_checkpoint_id_caching(
-        selected.note_commitment_tree_position(),
-        &anchor_height,
-    )?;
-
-    Ok(external_note
-        .filter(|n| expected_root == merkle_path.root(Node::from_cmu(&n.cmu())))
-        .map(|n| (n, Scope::External, merkle_path.clone()))
-        .or_else(|| {
-            internal_note
-                .filter(|n| expected_root == merkle_path.root(Node::from_cmu(&n.cmu())))
-                .map(|n| (n, Scope::Internal, merkle_path))
-        }))
 }